telegram-infrastructure/telegram-gateway/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
aiogram==3.*
nats-py
pydantic-settings
httpx
pyyaml
